+++ app/routes/toml-json.tsx
@@ -5,7 +5,7 @@ import { Button } from "~/components/ui/button";
 import { Textarea } from "~/components/ui/textarea";
 import { TipsCard } from "~/components/TipsCard";
 import { StatusAnnouncer } from "~/hooks/useStatusAnnouncement";
-import { useOutputCopy } from "~/hooks/useOutputCopy";
+import { useCopyWithFeedback } from "~/hooks/useCopyWithFeedback";
 import * as TOML from "smol-toml";
 
 export const Route = createFileRoute("/toml-json")({
@@ -92,7 +92,7 @@ export function jsonToToml(jsonStr: string): string {
  * TOML↔JSON相互変換コンポーネント
  */
 function TomlJsonConverter() {
-  const { statusRef, announceStatus, showToast, isCopied, handleCopy } = useOutputCopy();
+  const { statusRef, announceStatus, showToast, isCopied, handleCopy } = useCopyWithFeedback();
   const [mode, setMode] = useState<ConversionMode>("toml-to-json");
   const [indent, setIndent] = useState<2 | 4>(2);
   const [inputText, setInputText] = useState("");
